Descriptive statistics of Hydrogeochemical data
This code tries to implement data analytics on a dataset collected from the analysis of eight water samples.

The code only tries to implement descriptive statistical methods to inform decisions due to the size of the data. 

Branching should be done on the code repository for other code views to be accommodated.